requirements.txt
streamlit==1.31.0
openai==1.12.0
chromadb==0.4.22
pypdf2==3.0.1
python-dotenv==1.0.1
pydantic==2.6.1
aiohttp==3.9.3
tenacity==8.2.3
pandas==2.2.0